app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
sqlalchemy==2.0.43
aiosqlite==0.19.0
pydantic-settings==2.4.0
psycopg2-binary==2.9.10
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

bcrypt==4.0.1
redis==5.0.1
minio==7.1.16
prometheus-client==0.17.1 
python-dotenv==1.0.0
email-validator==2.1.0
aiosmtplib==2.0.0
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
freezegun==1.2.2
asynctest==0.13.0
websockets==12.0
Pillow==11.3.0
python-multipart==0.0.6
authlib==1.2.1
email-validator==2.1.0
python-json-logger==2.0.7
greenlet==3.2.4